feat: support conditional permissions
This commit passes the context and doc the permissions config, allowing permissions to change based on doc contents.
For example, the current user can be added to the modfiy permissions only when they are listed as the author of a document.
Use the context argument to get the current user's id.