fix: recreate collections after config change
When we deep copy collections, the function references in the collection objects are lost, but the property still exists, causing the `constructCollections` function to not actually construct the collection because it thinks it is already constructed. We now remove those proeprties before constructing the collections.
We also wipe out existing models so models based on the old config do not persist past the creation of the new config.