feat: port app ydoc utils features and add tests (#17)
* test: add tests for YBoolean

* test: add tests for YDate

* fix: provide document for fragment toDOM function

* test: add tests for number shared types

* fix: return tiptap editor json as json string instead of object

* test: add tests for YString

* fix: fall back to partially populated reference values

* fix: find docs by custom _id field

This now uses the custom field, making it match the process followed in other functions in YReference.

* fix: keep populated values in same order as unpopulated values

* feat: support including forced fields in reference

* feat: support requiring fields to exist

An example use case for this might be to filter to only allow a photo to be referenced when it has a photo credit field

* fix: only return partially populated fields if no reference collection is defined

* fix: allow dot notation for deep property keys

* test: add tests for YReference

* build(npm): add merge-anything

* feat: support geting doc array content with correct values

* fix: delete doc field shares on doc array delete

* fix: use unknown instead of any

* feat: add fuction to get fields from ydoc

* test: add tests for YDocArray

* docs(changeset): update ydoc utils to handle all cases and references

With the addition of tests for every ydoc util, setting and retrieving data becomes more predictible and reliable. Things that were previously supported in the web app are now supported in this package, including getting data from referenced collections (or not), predictable return types, and correct doc array values.